requirements.txt
gradio==3.50.2
Pillow==10.0.0